% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exset}
\alias{exset}
\title{Exact Gene-Based Testing}
\usage{
exset(statistic, ld_eigenvalues, iterations = 1e+05, type1_error = 0.05/12727)
}
\arguments{
\item{ld_eigenvalues}{eigenvalues of positive-definite LD matrix}

\item{iterations}{number of Monte Carlo replicates used to approximate integral}

\item{type1_error}{corrected type I error rate}
}
\description{
This function allows you to perform an exact hypothesis test of the gene-based null hypothesis
}
\examples{
exset()
}
\keyword{integration}
